+++ src/main/java/gift/controller/AdminController.java
@@ -1,11 +1,15 @@
-package gift;
+package gift.controller;
 
 
-import org.springframework.ui.Model;
+import gift.Product;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
-import java.util.*;
+import java.util.Collection;
 
 @Controller
 public class AdminController {
