+++ src/main/java/gift/service/ProductService.java
@@ -1,0 +1,4 @@
+package gift.service;
+
+public class ProductService {
+}
